requirements.txt
gradio>=6.6.0
python-dotenv>=0.9.9
edgartools[ai]>=5.13.0
openai>=2.16.0
openai-agents>=0.7.0
